+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldPointedEnchantment.java
@@ -1,0 +1,31 @@
+package com.insane96mcp.shieldsplus.world.item.enchantment;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ShieldItem;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+
+public class ShieldPointedEnchantment extends Enchantment {
+
+	public ShieldPointedEnchantment() {
+		super(Rarity.UNCOMMON, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
+	}
+
+	public int getMinCost(int p_44598_) {
+		return 12 + (p_44598_ - 1) * 20;
+	}
+
+	public int getMaxCost(int p_44600_) {
+		return this.getMinCost(p_44600_) + 25;
+	}
+
+	public int getMaxLevel() {
+		return 2;
+	}
+
+	@Override
+	public boolean canApplyAtEnchantingTable(ItemStack stack) {
+		return stack.getItem() instanceof ShieldItem;
+	}
+}